+++ controllers/messenger.js
@@ -6,7 +6,7 @@ const authHelper = require('../helpers/authentication');
 
 router.use(authHelper.authChecker);
 
-router.get('/token', (req, res, next) => api(req)
+router.get('/token', (req, res, next) => api(req, { timeout: 16000 })
 	.post('/messengerToken', {})
 	.then((result) => res.send(result))
 	.catch(next));