+++ helpers/apiHelper.js
@@ -6,8 +6,8 @@ if (Configuration.has('REQUEST_TIMEOUT_MS') !== true) {
 }
 
 const xApiKey = Configuration.get('API_KEY');
-const timeout = Configuration.get('REQUEST_TIMEOUT_MS');
-const api = (baseUrl, { keepAlive = false } = {}) => (req, { json = true } = {}) => {
+const defaultTimeout = Configuration.get('REQUEST_TIMEOUT_MS');
+const api = (baseUrl, { keepAlive = false, timeout = defaultTimeout } = {}) => (req, { json = true } = {}) => {
 	const headers = {};
 	if (req && req.cookies && req.cookies.jwt) {
 		headers.Authorization = (req.cookies.jwt.startsWith('Bearer ') ? '' : 'Bearer ') + req.cookies.jwt;